% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterMisc.R
\name{agg_fun}
\alias{agg_fun}
\title{Aggregate Raster}
\usage{
agg_fun(i, rast)
}
\arguments{
\item{i}{aggregation factor}

\item{rast}{raster being aggregated}
}
\value{
raster
}
\description{
Aggregate Raster
}
