% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHI.R
\name{count_in}
\alias{count_in}
\alias{count_in2}
\title{count number of neighbours locally}
\usage{
count_in(inraster)

count_in2(inraster)
}
\arguments{
\item{inraster}{}
}
\value{
a vector of length 1
}
\description{
count number of neighbours locally

More user-friendly but less-optimized version of count_in
}
