% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyCleaner.R
\name{COUNTIESm}
\alias{COUNTIESm}
\alias{COUNTIESmm}
\alias{COUNTIESh}
\title{Functions to read in County Shapefiles}
\usage{
COUNTIESm(myears, rddir)

COUNTIESmm(mmyears, rddir)

COUNTIESh(hyears, ddir)
}
\arguments{
\item{myears}{years to grab}

\item{rdir}{from what directory}
}
\value{
shapefile
}
\description{
Functions to read in County Shapefiles
}
\details{
Modern Counties

Middle Modern Counties

Historical Counties from 1690:2000
}
