% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterMisc.R
\name{CRAST_fun}
\alias{CRAST_fun}
\title{Transform Polygon Shapefile to Raster}
\usage{
CRAST_fun(SHP, field_name, Base, GDAL = TRUE, cropmask = TRUE)
}
\arguments{
\item{SHP}{spatialPolygonDF object}

\item{field_name}{which variable to turn into raster?}

\item{Base}{project SHP to this raster}

\item{GDAL}{which method}

\item{cropmask}{mask the raster afterwards}
}
\value{
raster
}
\description{
Transform Polygon Shapefile to Raster
}
