% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyCleaner.R
\name{create_counties}
\alias{create_counties}
\title{Wrapper Functions for County Shapefiles}
\usage{
create_counties(
  hyears = c(1980, 1990, 2000),
  mmyears = 2008:2010,
  myears = 2011:2015,
  ddir,
  rdir
)
}
\arguments{
\item{hyears, myears, mmyears}{vectors of years}

\item{ddir, rdir}{directory of shapefiles}
}
\value{
list of county shapefiles
}
\description{
Wrapper Functions for County Shapefiles
}
