% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF2Raster.R
\name{DF2Raster}
\alias{DF2Raster}
\alias{df2raster}
\title{Formatting dataframe as rasterstack}
\usage{
DF2Raster(DF)

df2raster(DF)
}
\arguments{
\item{DF}{dataframe}
}
\value{
rasterstack
}
\description{
Formatting dataframe as rasterstack
}
\seealso{
DF2stack
}
