% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF2Raster.R
\name{DF2stack}
\alias{DF2stack}
\title{Formatting dataframe as rasterstack}
\usage{
DF2stack(DF, dfname)
}
\arguments{
\item{DF}{dataframe}

\item{dfname}{names to keep}
}
\value{
rasterstack
}
\description{
Formatting dataframe as rasterstack
}
\seealso{
DF2Raster
}
