% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadGeography.R
\name{download.raster}
\alias{download.raster}
\title{Download Rasters}
\usage{
download.raster(shape_url, layer, outdir = getwd(), layer_new = layer)
}
\arguments{
\item{shape_url}{the directory containing the shape files (.shp, .shx, ...)}

\item{layer}{the name of file to download}

\item{outdir}{the directory where to save the files}

\item{layer_new}{the filename to save}
}
\description{
Download Rasters
}
