% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYADmaker_new.R
\name{dyad.maker0}
\alias{dyad.maker0}
\title{Make Skeleton for Dyadic Panel}
\usage{
dyad.maker0(times, dyad_name = c("Dyad", "Year"))
}
\arguments{
\item{times}{time period}

\item{dyad_name}{}
}
\value{
An empty list to be filled in dyad.maker1
}
\description{
Make Skeleton for Dyadic Panel
}
