% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYADmaker_new.R
\name{dyad.maker1}
\alias{dyad.maker1}
\title{Make A list of dyads}
\usage{
dyad.maker1(dyad, d_times, t_span, d_df, d_df_id, d_df_names, d_df2,
  d_df2_names, d_df2_aggnames1, d_df2_aggnames2, d_tab = NULL, d_tabx, d_taby)
}
\arguments{
\item{dyad}{, empty list from DYADmaker0 to be filled}

\item{d_times}{, time periods to create dyads}

\item{t_span}{number of periods after d_times to include for data-grouping}

\item{d_df}{, X variables to merge ()}

\item{d_df_id}{, X merger ID}

\item{d_df_names}{, X var name of ID}

\item{d_df2}{<- DT}

\item{d_df2_names}{<- "Start"}

\item{d_tab}{, Y variable to merge ( table of battles )}

\item{d_tabx}{, "BTABx"}

\item{d_taby}{, "BTABy"}

\item{d_df_aggnames1}{, X merger name}

\item{d_df_aggnames2}{<- X merger name}
}
\value{
A list
}
\description{
Make A list of dyads
}
