% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYADmaker.R
\name{DYADmaker0}
\alias{DYADmaker0}
\title{Make Skeleton for Dyadic Panel}
\usage{
DYADmaker0(times, dyad_name = c("Political", "Dyad", "Year"))
}
\arguments{
\item{times}{time periods}

\item{dyad_name}{}
}
\value{
An empty list to be filled in DYADmaker1
}
\description{
Make Skeleton for Dyadic Panel
}
