% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYADmaker.R
\name{DYADmaker2}
\alias{DYADmaker2}
\title{Dyad List Formatting}
\usage{
DYADmaker2(dfname, dyad, ...)
}
\arguments{
\item{dyad}{, DYAD}

\item{dfname=c("Polis1", }{"Polis2", "Battles", "Pol1.Politic", "Pol2.Politic", "Year")}
}
\value{
A dataframe with Battle and Political Data
}
\description{
Dyad List Formatting
}
\details{
Transforming List into Dyadic DF (Part 3)
}
