% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmallPolys.R
\name{getSmallPolys}
\alias{getSmallPolys}
\title{Trim Polygon of Small Areas}
\usage{
getSmallPolys(poly, minarea = 0.01)
}
\arguments{
\item{poly}{shapefile}

\item{minarea}{only get polygons > minarea}
}
\value{
shapefile
}
\description{
Trim Polygon of Small Areas
}
