% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHI.R
\name{HHI}
\alias{HHI}
\title{Calculate HHI for each raster cell}
\usage{
HHI(
  rast,
  wind,
  FUN = hhi_in,
  mask = NA,
  writedir = NA,
  HHIname = paste0("HHIrast_", wind[1], "_", wind[2])
)
}
\arguments{
\item{wind}{size of local windows to consider}

\item{FUN}{what to calculate}

\item{mask}{mask the values afterwards}

\item{writedir}{write the raster to hard disk}

\item{HHIname}{what to name the raster}

\item{Crast}{raster from which to perform calculations}
}
\value{
a raster
}
\description{
Calculate HHI for each raster cell
}
\examples{
 
}
