% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHI.R
\name{hhi_in}
\alias{hhi_in}
\alias{hhi_in2}
\title{calculate HHI locally}
\usage{
hhi_in(inraster)

hhi_in2(inraster)
}
\arguments{
\item{inraster}{}
}
\value{
a vector of length 1
}
\description{
calculate HHI locally

More user-friendly but less-optimized version of hhi_in
}
\examples{
rast <- raster::raster( matrix(runif(100), 10, 10) ) 
inrast <- as.integer(rast> .2)
hhi_rast <- raster::focal(inrast, w=matrix(1,3,3), hhi_in2)
}
