% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raster2DF.R
\name{layer_list}
\alias{layer_list}
\title{Transform RasterStack to Data.Table using Parallel Processing}
\usage{
layer_list(stack)
}
\arguments{
\item{stack}{stack of rasters to be converted, must have coordinate columns (x,y)}
}
\value{
datatable
}
\description{
Transform RasterStack to Data.Table using Parallel Processing
}
