% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYADmaker.R
\name{DYADmaker0}
\alias{DYADmaker0}
\title{Make Dyad Skeleton}
\usage{
DYADmaker0(times, dyad_name = c("Political", "Dyad", "Year"))
}
\arguments{
\item{times}{}
}
\value{
An empty list to be filled in DYADmaker1
}
\description{
Make Dyad Skeleton
}
